requirements.txt
Flask==3.1.1
Flask-OpenID==1.3.1
Flask-SQLAlchemy==3.1.1
launchpadlib==2.1.0
python-slugify==8.0.4
